+++ ui/mods/donation_panel/feed.js
@@ -1,6 +1,7 @@
 define([
   'donation_panel/tiltify/local',
   'donation_panel/tiltify/api_test',
+  'donation_panel/tiltify/live',
   'donation_panel/fundrazr/test',
   'donation_panel/fundrazr/live',
   'donation_panel/gofundme/test',
@@ -8,6 +9,7 @@ define([
 ], function(
   tiltify_local,
   tiltify_api_test,
+  tiltify_live,
   fundrazr_test,
   fundrazr_live,
   gofundme_test,
@@ -16,6 +18,7 @@ define([
   return {
     tiltify_local: tiltify_local,
     tiltify_api_test: tiltify_api_test,
+    tiltify_live: tiltify_live,
     fundrazr_test: fundrazr_test,
     fundrazr_live: fundrazr_live,
     gofundme_test: gofundme_test,